fix debugger crash
fusion seems to ask for debugger item immediately when you add it to debugger, but the render is delayed to display routine for sake of performance (unless you call render action to force it. expressions that involves render result will also call render before returns.). at that time object is not initialized, which causes crash.